requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Market Data
yfinance>=0.2.28
nsepy>=0.8
nsetools>=1.0.11
jugaad-trader>=0.26

# Technical Analysis
ta>=0.10.2
pandas-ta>=0.3.14b
scipy>=1.11.0

# NLP / News Sentiment
transformers>=4.35.0
torch>=2.0.0
newsapi-python>=0.2.7
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.3
newspaper3k>=0.2.8
vaderSentiment>=3.3.2

# Alerts
python-telegram-bot>=20.6
smtplib2>=0.2.1

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.1.0

# Scheduling
apscheduler>=3.10.4
schedule>=1.2.1

# Utilities
loguru>=0.7.2
rich>=13.6.0
tqdm>=4.66.1
pytz>=2023.3
httpx>=0.25.0
aiohttp>=3.9.0

# Storage
sqlalchemy>=2.0.0
